refactor: splitting more things
